fix(l1): satisfy external analyzer gates
